# swagger.json
swagger: '2.0'
info:
  title: Dispatcharr API
  description: API documentation for Dispatcharr
  termsOfService: https://www.google.com/policies/terms/
  contact:
    email: contact@dispatcharr.local
  license:
    name: Creative Commons by-nc-sa
  version: v1
host: 172.16.0.19:9191
schemes:
- http
basePath: /
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
- Bearer: []
paths:
...
  /api/channels/channels/:
    get:
      operationId: api_channels_channels_list
      description: ''
      parameters:
      - name: name
        in: query
        description: ''
        required: false
        type: string
      - name: channel_group
        in: query
        description: ''
        required: false
        type: string
      - name: epg
        in: query
        description: ''
        required: false
        type: string
      - name: search
        in: query
        description: A search term.
        required: false
        type: string
      - name: ordering
        in: query
        description: Which field to use when ordering the results.
        required: false
        type: string
      - name: page
        in: query
        description: A page number within the paginated result set.
        required: false
        type: integer
      - name: page_size
        in: query
        description: Number of results to return per page.
        required: false
        type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
            - count
            - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/Channel'
      tags:
      - api
    post:
      operationId: api_channels_channels_create
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/Channel'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/Channel'
      tags:
      - api
    parameters: []
...
  /api/channels/channels/{channel_id}/streams/:
    get:
      operationId: api_channels_channels_streams_list
      description: ''
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - api
    parameters:
    - name: channel_id
      in: path
      required: true
      type: string
  /api/channels/channels/{id}/:
    get:
      operationId: api_channels_channels_read
      description: ''
      parameters: []
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/Channel'
      tags:
      - api
    put:
      operationId: api_channels_channels_update
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/Channel'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/Channel'
      tags:
      - api
    patch:
      operationId: api_channels_channels_partial_update
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/Channel'
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/Channel'
      tags:
      - api
    delete:
      operationId: api_channels_channels_delete
      description: ''
      parameters: []
      responses:
        '204':
          description: ''
      tags:
      - api
    parameters:
    - name: id
      in: path
      description: A unique integer value identifying this channel.
      required: true
      type: integer
...
  /api/channels/streams/:
    get:
      operationId: api_channels_streams_list
      description: ''
      parameters:
      - name: name
        in: query
        description: ''
        required: false
        type: string
      - name: channel_group_name
        in: query
        description: ''
        required: false
        type: string
      - name: m3u_account
        in: query
        description: ''
        required: false
        type: string
      - name: m3u_account_name
        in: query
        description: ''
        required: false
        type: string
      - name: m3u_account_is_active
        in: query
        description: ''
        required: false
        type: string
      - name: search
        in: query
        description: A search term.
        required: false
        type: string
      - name: ordering
        in: query
        description: Which field to use when ordering the results.
        required: false
        type: string
      - name: page
        in: query
        description: A page number within the paginated result set.
        required: false
        type: integer
      - name: page_size
        in: query
        description: Number of results to return per page.
        required: false
        type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
            - count
            - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/Stream'
      tags:
      - api
    post:
      operationId: api_channels_streams_create
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/Stream'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/Stream'
      tags:
      - api
    parameters: []
...
definitions:
...
  Channel:
    required:
    - name
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      channel_number:
        title: Channel number
        type: number
        x-nullable: true
      name:
        title: Name
        type: string
        maxLength: 255
        minLength: 1
      channel_group_id:
        title: Channel group id
        type: integer
      tvg_id:
        title: Tvg id
        type: string
        maxLength: 255
        x-nullable: true
      tvc_guide_stationid:
        title: Tvc guide stationid
        type: string
        maxLength: 255
        x-nullable: true
      epg_data_id:
        title: Epg data id
        type: integer
        x-nullable: true
      streams:
        type: array
        items:
          type: integer
        uniqueItems: true
      stream_profile_id:
        title: Stream profile id
        type: integer
        x-nullable: true
      uuid:
        title: Uuid
        type: string
        format: uuid
        readOnly: true
      logo_id:
        title: Logo id
        type: integer
        x-nullable: true
      user_level:
        title: User level
        type: integer
        maximum: 2147483647
        minimum: -2147483648
      auto_created:
        title: Auto created
        description: Whether this channel was automatically created via M3U auto channel
          sync
        type: boolean
      auto_created_by:
        title: Auto created by
        description: The M3U account that auto-created this channel
        type: integer
        x-nullable: true
      auto_created_by_name:
        title: Auto created by name
        type: string
        readOnly: true
...
  Stream:
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      name:
        title: Name
        type: string
        maxLength: 255
        minLength: 1
      url:
        title: Url
        type: string
        x-nullable: true
      m3u_account:
        title: M3u account
        type: integer
        x-nullable: true
      logo_url:
        title: Logo url
        type: string
        x-nullable: true
      tvg_id:
        title: Tvg id
        type: string
        maxLength: 255
        x-nullable: true
      local_file:
        title: Local file
        type: string
        readOnly: true
        x-nullable: true
        format: uri
      current_viewers:
        title: Current viewers
        type: integer
        maximum: 2147483647
        minimum: 0
      updated_at:
        title: Updated at
        type: string
        format: date-time
        readOnly: true
      last_seen:
        title: Last seen
        type: string
        format: date-time
      stream_profile_id:
        title: Stream profile id
        type: integer
        x-nullable: true
      is_custom:
        title: Is custom
        description: Whether this is a user-created stream or from an M3U account
        type: boolean
      channel_group:
        title: Channel group
        type: integer
        x-nullable: true
      stream_hash:
        title: Stream hash
        description: Unique hash for this stream from the M3U account
        type: string
        maxLength: 255
        minLength: 1
        x-nullable: true
      stream_stats:
        title: Stream stats
        description: JSON object containing stream statistics like video codec, resolution,
          etc.
        type: object
        x-nullable: true
      stream_stats_updated_at:
        title: Stream stats updated at
        description: When stream statistics were last updated
        type: string
        format: date-time
        x-nullable: true
...